all Screen has been Move to Screen Directory and change the name of model directory
